[LowerToHW] Treat instance under bind as non-DUT
Change LowerToHW to not count bound instances as being under the DUT.
This is done as a stopgap measure to align FIRRTL layers with the
existing, legacy directory structure that CIRCT uses when provided with
both a MarkDUTAnnotation and a TestBenchDirAnnotation. Namely, anything
instantiated by the DUT is put in the default output directory and
anything instantiated only by other users (Grand Central, Test Harness,
etc.) is allowed to move into one of the directories that those other
users handle.  With this change, this will no longer cause layer-derived
modules or modules instantiated under layers exclusively to be moved into
the main output directory.

This is intended to be replaced with filelists/manifests where the build
system driving CIRCT is responsible for moving things into separate
directories after CIRCT finishes.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>