[ImportVerilog] Add type conversion and basic variables (#6755)
Implement the conversion from SystemVerilog types in Slang's AST to the
corresponding MLIR types in the Moore dialect. To facilitate testing,
add a basic `moore.variable` op and convert module-level variable
declarations to this new op. This requires converting the variable's
type. We may choose to have variables return a special lvalue or pointer
wrapper type in the future once such a need arises to clearly demarcate
variable reading and assignment.

Since we don't support expressions yet, variables with an initializer
expression cause an error for the time being.