[ci] Allow uploading wheels from workflow_dispatch
Add an input to the workflow_dispatch even of the "Upload Wheels" GitHub
Action workflow.  This is done to enable a user to build the wheels and
upload them without having to wait for a weekly release to do this or
having to ask somebody with the keys to do the upload for them.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>