[FIRRTL] Handle reference ports when Classes dedup. (#6770)
In https://github.com/llvm/circt/pull/6582, initial support for
classes and objects was added in Dedup. However, since classes and
objects are new constructs, not every possibility was handled in the
initial implementation.

This specifically handles the case where references to objects are
passed through class ports, and the class type of such objects has
changed because their classes deduped.

In the final fixup pass through the instance graph, if we find a
class, we check if it has any reference ports that need to be updated,
and if so update them. When this happens, we also update an objects of
the class to reflect the newly updated class type.

Fixes https://github.com/llvm/circt/issues/6603.