[CFToHandshake] Don't modifyOpInPlace in PartialLowerOp.
The modifyOpInPlace seems to have been used to give the
PartialLoweringFunc carte blanche to make in-place modifications
without using the ConversionPatternRewriter. However, the one and only
user of this ConversionPattern actually erases the source op and
performs no in-place modifications.

The DialectConversion infrastructure now asserts that erased ops are
not modified, so wrapping a function that erases an op in
modifyOpInPlace will assert.

This simply remove the unnecessary use of modifyOpInPlace, and updates
the comment about the required usage.