[FIRRTL] Handle either kind of path attribute in LowerClasses.
Sometimes the start of a hierarchical path is an InnerRefAttr, and
sometimes the path is just a single FlatSymbolRefAttr. Either way, we
just need the StringAttr refering to the module name. There is a
helper called `root` for this on HierPathOp. But at this point we are
constructing the path to build a HierPathOp, we don't already have
one. So, I've implemented this the same as how `root` is implemented.