[ImportVerilog] Add if and loop statements (#6831)
Extend the `ImportVerilog` conversion to support `if` statements and
`for`, `repeat`, `while`, `do ... while`, and `forever` loops. Lower
them to `scf.if` and `scf.while` ops for now. In the future, since
Verilog has `break`, `continue`, and `return` statements, we may have to
switch to a different dialect or defined these ops in the Moore dialect.

Co-authored-by: Hailong Sun <hailong.sun@terapines.com>
Co-authored-by: ShiZuoye <albertethon@163.com>
Co-authored-by: hunterzju <hunter_ht@zju.edu.cn>
Co-authored-by: Anqi Yu <anqi.yu@terapines.com>