[InstanceGraph] Expose choice of top in getAbsolutePaths, NFC.
This exposes a new getAbsolutePaths API where the user may specify an
InstanceGraphNode to use as the top of the returned paths. The
existing API that does not expose this simply calls the new API using
getTopLevelNode(), preserving the previous behavior. An upcoming
change will use the new API to get absolute paths relative to a
different top.