[NFC][clang-tidy] Disallow global 'using' directives in headers (#6998)
Add the google-global-names-in-headers rule to the list of clang-tidy checks to prevent namespace pollution.

This is intended to flag global using-directives in header files during PR checks. Collisions between the llvm and the mlir namespace have caused obscure build failures on Windows in the past. #6844 should have removed all existing occurrences in the CIRCT code base.