[Python] Remove dependency on MLIRPythonSources.Dialects.builtin.
This specific dependency no longer works after an upstream change:
https://github.com/llvm/llvm-project/pull/97167. Furthermore, we never
actually wanted this dependency, since we don't actually do anything
with the builtin dialect ops. We only used this to convince the MLIR
bindings to do the automatic downcasting for builtin dialect types and
attributes, which required the builtin dialect to be registered (even
though it always is) until this change went in:
https://github.com/llvm/llvm-project/pull/72338. Now that these
changes are behind us, we can remove our workaround and avoid the
issues that ensue by depending on a target no one usually depends on.