[FIRRTL][ProbesToSignals] Add pass to replace probes with signals. (#7342)
* [FIRRTL][ProbesToSignals] Add pass to replace probes with signals.

Fixes #7317.

The noRefTypePorts option is still needed in practice until a better
story for memtaps is developed.  However, since those are only
introduced by annotations that option can continue to be used
for those while this pass addresses what's reachable from
FIRRTL inputs themselves (and not by that option).

The pass nevertheless does include the functionality needed to
convert the CHIRRTL debug ports that mem taps produce into what
that option produces instead.  However it does not presently
support what those are lowered to (and its forms through the pipeline).