[FIRRTL] Remove module self-instantiation verifier
This removes a verifier that checks that modules do not instantiate
themsselves.  This will be covered by a new diagnostic pass which can
handle indirect recursion as well.  This instance verifier goes a bit
beyond what we normally check in verifiers, as it must reach upward to
find the owning module to compare against its target module, and we
normally try to check only local operation properties in verifiers.