[FIRRTL] Cache symbol table in LowerLayers (#7436)
Precompute a symbol to layer mapping inside LowerLayers and uses this
instead of a symbol table.  This is both faster and avoids problems of
trying to compute a symbol table while modules may be created.

Repurpose the functions that were being used to create macro declarations
to also compute this symbol to layer mapping.  Rename these to indicate
that they are now doing generic layer preprocessing.

Fixes #7434.

h/t @youngar for the fix suggestion.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>