[FIRRTL] Fix sub-* op in layer block verifier
Fix issues with verification of subfield, subindex, and subaccess
operations which appear in a layer block.  These operations are allowed to
occur in a layer block even if they capture non-passive operands.

This requires reworking layer block verification to no longer check for
operations using non-passive operands.  The spec requires that no
operation in a layer block _drives_ a value declared outside the layer
block.  However, this is exceedingly difficult to verify due to the fact
that non-passive destinations in ConnectLike operations can be
source-to-destination, destination-to-source, or bi-directional.  If the
verifier sees this, just allow it.  The FIRRTL pass pipeline will later
canonicalize away flips (i.e., make all types passive) which will then
allow the verifier to check these.  This should be revisited in the
future.

Fixes #7451.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

Don't check non-passive connects.