[FIRRTL] Compute LCA of BlackBoxDir output files
Change how the output directory is computed for external modules with
black box annotations.  Previously, this relied on a "precedence" of
predefined directories.  Now, this trivially computes the LCA of the
directories for all output directories for a black box annotation with the
same name field.

This _does not_ replicate the old behavior.  Instead, users are expected
to rework their output directory structure to align with this algorithm.
E.g., it is no longer possible to make the tesbench a sibling directory of
the main output directory and have blackboxes that are instantiated by
both the testbench and the design to be put in the design---instead a user
should nest the testbench directory _under_ the main output directory.

The omnibus test case has been necessarily updated to show how the above
nesting.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>