[FreezePaths] Add an optional argument, to get the operation name (#7549)
`FreezePaths` pass may sometimes need to be invoked before `ExportVerilog`. 
This change removes the dependency of the pass on the `hw.verilogName`
 attribute, and adds an optional get operation name function, that can be
 invoked if the verilog name is absent. 
This enables any tool to invoke it earlier by providing an appropriate 
name-getter function. 
But as is expected, the names used to freeze the paths, may not match the
 verilog names. This should be used with caution, only when the path will not
 be used and its okay to be incorrect.