[FIRRTL] Add "effective" functions to InstanceInfo
Add two new functions for computing if a module is under the "effective"
design-under-test (DUT):

  - atLeastOneInstanceUnderEffectiveDut
  - allInstancesUnderEffectiveDut

These functions are very simple.  However, their use comes up in a number
of passes and this makes _reading_ those passes more straightforward.
I.e., it's quicker to grok what is going on with these functions than with
`!hasDut() || atLeastOneInstanceUnderEffectiveDut()`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>