[OM] Update FreezePaths to handle object fields with paths. (#7683)
When we walk the IR and freeze paths, we previously didn't handle
object field ops. This is totally valid IR, and is another case where
we may encounter ops defining SSA values of path types that need to be
frozen.

This adds a visitor for ObjectFieldOp, similar to the other visitors,
which rewrites the result type to map any path types to frozen path
types.