[FIRRTL] InferWidths: don't store all exprs in a vector
We are currently storing all allocated expressions in a single giant
vector, which we later use to find VarExprs and DerivedExprs.  This
changes it to keep two separate dedicated vectors which only store the
type of exprs we are interested in. It would be nice if we could iterate
over all objects in a SpecificBumpPtrAllocator, but they don't expose
such an API.