[FIRRTL] Create DUT Metadata if it exists
Fix a bug in the `CreateSiFiveMetadata` pass where metadata about the
design-under-test (DUT) would not be created unless blackbox, memory, or
retimed modules were also present.  While this is likely a bug fix (of
behavior that is admittedly not defined anywhere), it is primarily
motivated from a desire to make the tests of this pass more modular.
I.e., I want to change these tests to test only DUT metadata emission.
This fix/change is in direct support of that aim.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>