[FIRRTL] use properties in builders and parsers (#7766)
This change makes op builders and parser populate the property struct
directly instead of adding them to the attr-dict.  During operation
creating, any inherent attributes are copied out of the
result.attributes and in to the properties struct.  It is more idiomatic
and slightly more efficient to store inherent attributes directly in
properties.