[FIRRTL] Simplify CreateSiFiveMetadata tests, NFC
Add additional, smaller tests of the `CreateSiFiveMetadata` pass related
to how it handles metadata emission for corner cases.  E.g., when there
are memories that are instantiated both in the testharness and the
design-under-test (DUT).  This is work in preparation to modify this pass
to handle layers.

Remove the omnibus test of everything (which was only spottily testing
certain features).  This can obviously continue to be improved over time.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>