[FIRRTL] Make metadata symbols private
For internal symbols created in the `CreateSiFiveMetadata` pass, make
these private and not public.  This caused some downstream errors in
an out-of-tree project which was doing MLIR-level linking and not renaming
duplicate symbols.  While this downstream project will be fixed, it is
also better to not have these symbols public as that prevents them from
being DCE'd.

h/t @uenoku for the proposed fix here.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>