[FIRRTL] Dedup: speed up handling of instances (#7815)
Dedup tries to hash all modules in parallel.  To accomplish this, the
names of instantiated modules are not included as part of the structural
hash, but they are taken in to account when checking if two modules are
the same.  This process involves comparing the instantiated children
modules of two modules if their hashes match.  This was implemented by
using an array attribute, to make comparisons quicker.

When a module or class has many thousands of instances underneath it, it
becomes impractical to build a array attribute with every child module.
Interning such a large ArrayAttr is incredibly slow and will eat up that
memory for the rest of the process.

Instead, we don't bother interning the instance arrays, and just keep
them as plain old vectors, which comes with the benefit of not eagerly
interning gigantic arrays.