[FIRRTL] Dedup: record less indices when hashing (#7816)
We map operations to their index as they appear in the IR, so that we
can use the index when hashing their uses as operands and inner-symbol
references. We only need to bother assigning indices to operations which
actually have results or an inner-symbol.  Technically we can do a
little better, the result or symbol has to actually have a user, but we
we don't have many trivially dead expressions on our input.