[LowerClasses] Stop cloning operations one at a time, NFCI.
The previous implementation would clone operations from the original
module-like body to the new OM class body one at a time, maintaining
an IRMapping as it went. This becomes a pretty bad performance
bottleneck as the IR size increases, and the DenseMap within IRMapping
starts to dominate.

Instead, this clones the entire region wholesale, and splices the body
into the new OM class op. It then traverses both the original
module-like and the OM class op to clean up each side: removing
property ops from the module-like, and removing non-property ops from
the OM class op.

This should be a non-functional change, but should dramatically help
with scalability of this pass.