[FIRRTL] Simplify, add GC to InstanceInfo
Simplify the tedious logic in the InstanceInfo analysis to more directly
push lattice values down the instance graph.  Previously, this uses
special casing during this walk to handle values that were undefined.
This new version sets all the values at the boundary of the graph (modules
which are never instantiated) and then only deals with a minimum of
special handling for determining whether or not user information is
merged.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>