[FIRRTL] Add GC Companion to InstanceInfo Analysis
Change the InstanceInfo analysis to record Grand Central Companions as
being _not_ in the design.  These are functionally the same as layers
except that this is represented as a module as opposed to an operation
with a block.  This is done as part of updating passes to use the
`InstanceInfo` analysis as opposed to relying on their own, individual
computations of what modules are in or not in the "design".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>