[FIRRTL] No annotations may be deleted
Change Annotation handling behavior so that no annotation can be deleted.
Previously, there existed a method that would report if an annotation
could be deleted.  However, now that the only annotation that had this
semantic, OMIRTracker, is gone, this method is now always `false` for a
single annotation and equivalent to `empty` for an `AnnotationSet`.  Use
these simpler functions instead and delete the `canBeDeleted` functions.

Co-authored-by: Mike Urbach <mike.urbach@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>