[FIRRTL] Support regions/blocks in LowerLayers
Fix a bug in the `LowerLayers` pass where it would not visit the blocks in
operations with nested regions.  This would cause the pass to produce
invalid IR if the operations within a layerblock contained any operations
which contained blocks which had captures outside the block.  E.g., this
prevented the pass from working correctly with when statements.

This is part of enabling inline layers.  The existing lowering of inline
layers produces `sv.ifdef` operations.  If an inline layer is inside a
bind layer, then the inline layer (lowered to an `sv.ifdef`) needs to be
walked to find any captures.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>