[FIRRTL] Drop DUT prefix code in ExtractInstances
Now that module prefixing support in CIRCT has been dropped, remove some
code which is now unreachable in the `ExtractInstances` pass.  This code
relied on the, now deleted, `PrefixModules` pass to set an undocumented
"prefix" field in the DUT annotation.  Now that this pass is gone, this
field will never exist.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>