[FIRRTL] Clock gate extraction work w/ prefixing (#7946)
Weaken the check used by the `ExtractInstances` pass to determine what is
a clock gate.  Change this from requiring that the defname of an external
module is exactly "EICG_wrapper" to instead only needing to end with
"EICG_wrapper".  This is done so that this will work with Chisel-time
prefixing and that we would actually _like_ to prefix these clock gates.

Commentary: this string matching is terrible.  I realize that.  However,
this is a legacy API that is intending to be replaced.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>