[ExportVerilog] Fix ifdef of macro w/ Verilog name
Fix a bug in the emission of `sv.ifdef` (and its procedural variant) where
the symbol name was always used instead of using the symbol name only if a
Verilog name was not specified.

This is done to fix an issue related to lowering of FIRRTL inline layers
where the `LowerLayers` pass generates new `sv.macro.decl` symbols that
will replace existing `firrtl.layer` symbols.  However, it relies on a
namespace to generate these and will result in suffixed symbol names.  (It
could obviously do more work to conserve symbols or generate in-pass
invalid IR with duplicate symbols.)  Nonetheless, the real bug here is
that the `ExportVerilog` conversion is not doing the right thing with the
Verilog name attribute.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>