[FIRRTL] Align inline layer ABI w/ spec proposal
Partially implement the inline layer convention ABI as described in the
FIRRTL Spec/ABI PR for this feature [[1]].  This changes the name of the
macro to be something that is more descriptive and cannot collide with
FIRRTL identifiers (which are supposed to not include `$`).

This does not implement the full ABI.  Namely, it does not correctly
handle circuits with multiple public modules.  (Notably, this isn't
implemented for the bind layer convention either.)  This will be fixed in
a follow-on.

[1]: https://github.com/chipsalliance/firrtl-spec/pull/262

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>