[Python] Remove multiple inheritance from Comb helper class.
As we move towards nanobind, the main issue is that nanobind
explicitly does not want to participate in multiple inheritance, which
means we can't use multiple inheritance in OpViews. For the comparator
comb ops, we decorate helper classes for conctructs like EqOp that
delegate to ICmpOp. The decorator creates a helper class that extends
OpView with multiple inheritance.

We can avoid the multiple inheritance with minor duplication by having
each concrete leaf class inherit OpView, rather than mixing in OpView
in the generated helper class.