[Python] Remove multiple inheritance from Seq helper class.
As we move towards nanobind, the main issue is that nanobind
explicitly does not want to participate in multiple inheritance, which
means we can't use multiple inheritance in OpViews. In the Seq
bindings, we mix in the CompRegLike class with multiple inheritance to
share the constructor code betwee CompRegOp and
CompRegClockEnabledOp.

To avoid multiple inheritance with a little duplication, this changes
the __init__ method on CompRegLike to a staticmethod, and adds
corresponding __init__ methods to CompRegOp and CompRegClockEnabledOp
that delegate to the implementation defined on CompRegLike.