[FIRRTL] Use InstanceInfo in BlackBoxReader
Convert the BlackBoxReader pass to use the InstanceInfo analysis as
opposed to relying on manually computed "isDut" information.

This is a slight change to the pass!  Previously, BlackBoxReader would
_not_ consider layers when determining if something is in the design.  It
would only consider the `lowerToBind` attribute.  The change in this
commit makes it such that blackboxes under layers will not be treated as
in the design.  This made no changes to the relatively exhaustive tests of
this, however, this was discovered when doing a direct comparison of the
original `isDut` member function to what the InstanceInfo analysis member
functions return.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>