[circt-test] Keep a list of available test runners (#8046)
Instead of accepting a runner binary directly through the `-r` command
line option, add a `RunnerSuite` to circt-test which tracks a list of
configured runners. This currently gets populated with the default "sby"
and "circt-bmc" runners with the corresponding configuration. The user
can then pass the runner name to the `-r` option, e.g. `-r sby`, instead
of a full path.

Further down the road we'll want to allow the user to dynamically
configure additional runners through files on disk, and specify an order
in which the runners should be used. For example, a user may want to use
"sby" for as many tests as possible, and fall back to "circt-bmc" for
the ones where "sby" was not available.