[FIRRTL] FART: allow modules to be in multiple domains at once (#8172)
When a child module is underneath two different reset domains, the pass
would emit an error message.  This change makes it so that it is allowed
as long as the reset signal name and the reset signal type between the
two domains are the same.

This also changes the logic to re-use a preexisting reset port: if there
is a port with the correct name for a reset domain, but the type does
match, we used to pick a different port name.  This changes it to be an
error, which matches the old SFC behaviour.