[ESI] Fix indirect MMIO and allocate more space
The initial implementation of MMIO indirection was broken. It would
produce many writes into the virtual MMIO space for every write to the
MMIO register. Works fine for idempotent registers, not so much for
side-effectful registers. Replace it with an implementation based on
channels. Not tested in cosim, only used by the XRT BSP.

Also restore the large address space handed out by the MMIO service
since we now have a 32-bit address space on all the supported platforms.