[LowerToHW] Lower PrintfOp w/ TimeOp
Add a lowering for PrintfOps which use TimeOps for substitutions.  A
TimeOp used by a PrintfOp will be lowered to an sv::TimeOp that is
co-located with the Verilog FWriteOp.  (This helps to avoid problems where
the system function `$time` could get spilled to a wire.  This is sunk
into the same procedural region as the fwrite to avoid this.)

Practically, this is the last piece needed to compile FIRRTL that includes
the `{{SimulationTime}}` special substitution all the way to Verilog.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>