[FIRRTL] Lower firrtl.simulation to verif.simulation (#8312)
Add a lowering from `firrtl.simulation` to `verif.simulation` in the
FIRRTL-to-HW conversion. The lowering creates a `verif.simulation` op
that simply instantiates the module pointed to by `firrtl.simulation`,
connecting its `clock` and `init` ports to the simulation op's block
arguments, and yielding the `done` and `success` ports back. We may want
to inline the instance in the future, but that is purely cosmetic.

This now allows FIR files to contain `simulation` unit tests that can be
lowered to HW through `firtool` and then executed using `arcilator`.