[FIRRTL] Add a bind op (#8384)
Currently, FIRRTL instances may be marked as "lowerToBind", which is then
lowered to a HW instance marked "doNotPrint", along with a bind op in the
global scope.

This PR adds a bind op to FIRRTL, which in conjunction with the "doNotPrint"
flag on FIRRTL instances, allows us to choose precisely where the bind
operation goes, rather than having lowerToHW choose for us.