[FIRRTL] Do not guard fprintf with PRINTF_COND
Remove the guard of fprintf from beign gated by PRINTF_COND.  While we may
want to eventually turn this back on, this is easier to change in the
future.  Additionally, there are much _better_ Chisel APIs to support this
now, e.g., the disable set of APIs.

Pragmatically, this is done because internally there most simulations have
printfs turned off and gated by a known test reset _and a plusarg_.
Hence, we're having problems landing this internally.  That said, I do
think that PRINTF_COND was _originally_ a mistake and I think we can
sidestep this for any _new_ API.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>