[ci] Use Ubuntu 24.04 for artifact publishing
Change the Ubuntu versioned used for native artifact publishing from
Ubuntu 22.04 to 24.04.

This is done because of internal consumers that use the 24.04 self-hosted
runners and 24.04 cannot consume 22.04 installs due to GNU libc
compatibility.

While I can see a benefit in providing the lowest common denominator
builds, the CIRCT shared/"static" full installs are basically best effort.
Anybody doing anything serious with this will very likely need to build
from source.  Alternatively, these are really only intended to be used by
GitHub Actions CI and need to be compatible there.  These published
installs have a very small user base and are a poor substitute for true
"packaging".  I.e., if we want to take this seriously, we need to be
building deb/rpm packages for specific versions and not just publishing an
install from some Linux distribution.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>