[OM] Add field_locs array attribute for ClassFieldsOp locations (#8439)
This changes the ClassFieldsOp logic to use an inherent attribute to store an array mapping field to location by index. This avoids the issue where FusedLoc merges non-unique locations (preventing us from retrieving a fields location by mapping it to an index). It also avoids an issue where metadata is removed from the FusedLoc that was previously being used to store this information.