[RTG][Elaboration] Arith support (#8366)
Initially the idea was to use the index dialect for arithmetic because it is smaller and simpler (doesn't accept any kind of shape types that we have to error out for, etc.). However, the SCF dialect depends on arith and its canonicalizers produce arith ops. The easiest solution to properly support if-statements and for loops in RTG is thus to just add support for arith ops that can be produced by those canonicalizers to the elaboration pass.